refactor(backend): 디렉토리 명명 규칙 복수형으로 통일
- const → consts
- guard → guards
- type → types
- pipe → pipes
