+++ apps/backend/src/auth/auth.module.ts
@@ -3,7 +3,7 @@ import { JwtModule } from '@nestjs/jwt';
 import { UsersModule } from '@/users/users.module';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { TokenGuard } from './guard/token.guard';
+import { TokenGuard } from './guards/token.guard';
 
 @Module({
 	imports: [JwtModule.register({}), forwardRef(() => UsersModule)],
