+++ apps/backend/src/users/pipes/password.pipe.ts
